added comments for clarity
