+++ public/style.css
@@ -34,6 +34,7 @@ nav a {
 .join-form {
   width: 35em;
   margin: 0 auto;
+  font-family: Arial, Helvetica, sans-serif;
 }
 
 .name-label {
@@ -56,4 +57,12 @@ nav a {
   margin: 1em 16.5em;
   border: 0.125em blue solid;
   border-radius: 0.25em;
+}
+
+.create-room-btn:hover {
+  background-color: #ffffff;
+}
+
+.create-room-btn:active {
+  background-color: #e9ecea;
 }
